Rename files and directories to what I hope are more accurate names
