+++ src/components/get_related_content_resolver.py
